fix: hvn_cidr is now seen as optional input variable
having null as default value makes the input variable 'required' which is not the the intention.
when passing null (module's default value), hvn_cidr is automatically set by the provider (currently 172.25.16.0/20 on both AWS and Azure).
We want to preserve this behavior.

If needing to pass a specific value for hvn_cidr, it will take precedence